Erro mensagem Não permitir excluir Manual com vínculo

+++ src/main/java/br/com/basis/abaco/web/rest/ManualResource.java
@@ -172,6 +172,13 @@ public class ManualResource {
     @Timed
     public ResponseEntity<Void> deleteManual(@PathVariable Long id) {
         log.debug("REST request to delete Manual : {}", id);
+
+        if(manualRepository.quantidadeContrato(id) > 0) {
+            return ResponseEntity.badRequest().headers(
+                HeaderUtil.createFailureAlert(ENTITY_NAME, "contratoexists", "A manual cannot be deleted"))
+                .body(null);
+        }
+
         manualRepository.delete(id);
         manualSearchRepository.delete(id);
         return ResponseEntity.ok().headers(HeaderUtil.createEntityDeletionAlert(ENTITY_NAME, id.toString())).build();
